requirements.txt
flask
flask-cors
gunicorn

numpy
opencv-python-headless
pillow
pymupdf

torch
torchvision
ultralytics
